@@ -1,38 +1,61 @@
 {
-    "name": "robmorgan/phinx",
+    "name": "ingenerator/phinx-lightweight",
     "type": "library",
-    "description": "Phinx makes it ridiculously easy to manage the database migrations for your PHP app.",
-    "keywords": ["phinx", "migrations", "database", "db", "database migrations"],
-    "homepage": "https://phinx.org",
+    "description": "Very simple SQL-based database migrations tool - a heavily stripped down fork of robmorgan/phinx",
+    "keywords": [
+        "phinx",
+        "migrations",
+        "database",
+        "db",
+        "database migrations"
+    ],
+    "homepage": "https://github.com/ingenerator/phinx-lightweight",
     "license": "MIT",
-    "version": "0.9.2",
-    "authors": [{
-        "name": "Rob Morgan",
-        "email": "robbym@gmail.com",
-        "homepage": "https://robmorgan.id.au",
-        "role": "Lead Developer"
-    }, {
-        "name": "Woody Gilk",
-        "email": "woody.gilk@gmail.com",
-        "homepage": "http://shadowhand.me",
-        "role": "Developer"
-    }, {
-        "name": "Richard Quadling",
-        "email": "rquadling@gmail.com",
-        "role": "Developer"
-    }, {
-        "name": "CakePHP Community",
-        "homepage": "https://github.com/cakephp/phinx/graphs/contributors"
-    }],
+    "authors": [
+        {
+            "name": "Andrew Coulton",
+            "email": "andrew@ingenerator.com",
+            "role": "Developer"
+        },
+        {
+            "name": "Craig Gosman",
+            "email": "craig@ingenerator.com",
+            "role": "Developer"
+        },
+        {
+            "name": "Rob Morgan",
+            "email": "robbym@gmail.com",
+            "homepage": "https://robmorgan.id.au",
+            "role": "Original Lead Developer"
+        },
+        {
+            "name": "Woody Gilk",
+            "email": "woody.gilk@gmail.com",
+            "homepage": "http://shadowhand.me",
+            "role": "Original Developer"
+        },
+        {
+            "name": "Richard Quadling",
+            "email": "rquadling@gmail.com",
+            "role": "Original Developer"
+        },
+        {
+            "name": "CakePHP Community",
+            "homepage": "https://github.com/cakephp/phinx/graphs/contributors",
+            "role": "Original contributors"
+        }
+    ],
     "require": {
-        "php": ">=5.4",
-        "symfony/console": "^2.8|^3.0|^4.0",
-        "symfony/config": "^2.8|^3.0|^4.0",
-        "symfony/yaml": "^2.8|^3.0|^4.0"
+        "php": "~8.0.0 || ~8.1.0 || ~8.2.0",
+        "symfony/console": "^4.0",
+        "symfony/config": "^4.0",
+        "symfony/yaml": "^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.35|^5.7|^6.5",
-        "cakephp/cakephp-codesniffer": "^3.0"
+        "phpunit/phpunit": "^9.5.5"
+    },
+    "conflict": {
+        "robmorgan/phinx": "*"
     },
     "autoload": {
         "psr-4": {
@@ -44,14 +67,7 @@
             "Test\\Phinx\\": "tests/Phinx"
         }
     },
-    "scripts": {
-        "check": [
-            "@test",
-            "@cs-check"
-        ],
-        "cs-check": "phpcs --colors -p --standard=vendor/cakephp/cakephp-codesniffer/CakePHP ./src ./tests ./app",
-        "cs-fix": "phpcbf --colors --standard=vendor/cakephp/cakephp-codesniffer/CakePHP ./src ./tests ./app",
-        "test": "phpunit --colors=always"
-    },
-    "bin": ["bin/phinx"]
+    "bin": [
+        "bin/phinx"
+    ]
 }
