@@ -31,8 +31,6 @@ namespace Phinx\Migration;
 use Phinx\Config\ConfigInterface;
 use Phinx\Config\NamespaceAwareInterface;
 use Phinx\Migration\Manager\Environment;
-use Phinx\Seed\AbstractSeed;
-use Phinx\Seed\SeedInterface;
 use Phinx\Util\Util;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -63,11 +61,6 @@ class Manager
      * @var array
      */
     protected $migrations;
-
-    /**
-     * @var array
-     */
-    protected $seeds;
 
     /**
      * @var integer
@@ -197,14 +190,10 @@ class Manager
                     '%s %14.0f  %19s  %19s  <comment>%s</comment>',
                     $status,
                     $migration->getVersion(),
-                    $version['start_time'],
-                    $version['end_time'],
+                    $version ? $version['start_time'] : NULL,
+                    $version ? $version['end_time'] : NULL,
                     $migration->getName()
                 ));
-
-                if ($version && $version['breakpoint']) {
-                    $output->writeln('         <error>BREAKPOINT SET</error>');
-                }
 
                 $migrations[] = ['migration_status' => trim(strip_tags($status)), 'migration_id' => sprintf('%14.0f', $migration->getVersion()), 'migration_name' => $migration->getName()];
                 unset($versions[$migration->getVersion()]);
@@ -264,9 +253,6 @@ class Manager
             str_pad($version['migration_name'], $maxNameLength, ' ')
         ));
 
-        if ($version && $version['breakpoint']) {
-            $this->getOutput()->writeln('         <error>BREAKPOINT SET</error>');
-        }
     }
 
     /**
@@ -324,21 +310,14 @@ class Manager
             }
         }
 
-        // are we migrating up or down?
-        $direction = $version > $current ? MigrationInterface::UP : MigrationInterface::DOWN;
-
-        if ($direction === MigrationInterface::DOWN) {
-            // run downs first
-            krsort($migrations);
-            foreach ($migrations as $migration) {
-                if ($migration->getVersion() <= $version) {
-                    break;
-                }
-
-                if (in_array($migration->getVersion(), $versions)) {
-                    $this->executeMigration($environment, $migration, MigrationInterface::DOWN);
-                }
-            }
+        if ($version < $current) {
+            throw new \RuntimeException(
+                sprintf(
+                    'Cannot migrate down to an earlier version (requested %s, but we are at %s)',
+                    $version,
+                    $current
+                )
+            );
         }
 
         ksort($migrations);
@@ -348,7 +327,7 @@ class Manager
             }
 
             if (!in_array($migration->getVersion(), $versions)) {
-                $this->executeMigration($environment, $migration, MigrationInterface::UP);
+                $this->executeMigration($environment, $migration);
             }
         }
     }
@@ -356,198 +335,31 @@ class Manager
     /**
      * Execute a migration against the specified environment.
      *
-     * @param string $name Environment Name
+     * @param string                              $name      Environment Name
      * @param \Phinx\Migration\MigrationInterface $migration Migration
-     * @param string $direction Direction
+     *
      * @return void
      */
-    public function executeMigration($name, MigrationInterface $migration, $direction = MigrationInterface::UP)
+    public function executeMigration($name, MigrationInterface $migration)
     {
         $this->getOutput()->writeln('');
         $this->getOutput()->writeln(
             ' ==' .
             ' <info>' . $migration->getVersion() . ' ' . $migration->getName() . ':</info>' .
-            ' <comment>' . ($direction === MigrationInterface::UP ? 'migrating' : 'reverting') . '</comment>'
+            ' <comment>migrating</comment>'
         );
 
         // Execute the migration and log the time elapsed.
         $start = microtime(true);
-        $this->getEnvironment($name)->executeMigration($migration, $direction);
+        $this->getEnvironment($name)->executeMigration($migration);
         $end = microtime(true);
 
         $this->getOutput()->writeln(
             ' ==' .
             ' <info>' . $migration->getVersion() . ' ' . $migration->getName() . ':</info>' .
-            ' <comment>' . ($direction === MigrationInterface::UP ? 'migrated' : 'reverted') .
+            ' <comment>migrated' .
             ' ' . sprintf('%.4fs', $end - $start) . '</comment>'
         );
-    }
-
-    /**
-     * Execute a seeder against the specified environment.
-     *
-     * @param string $name Environment Name
-     * @param \Phinx\Seed\SeedInterface $seed Seed
-     * @return void
-     */
-    public function executeSeed($name, SeedInterface $seed)
-    {
-        $this->getOutput()->writeln('');
-        $this->getOutput()->writeln(
-            ' ==' .
-            ' <info>' . $seed->getName() . ':</info>' .
-            ' <comment>seeding</comment>'
-        );
-
-        // Execute the seeder and log the time elapsed.
-        $start = microtime(true);
-        $this->getEnvironment($name)->executeSeed($seed);
-        $end = microtime(true);
-
-        $this->getOutput()->writeln(
-            ' ==' .
-            ' <info>' . $seed->getName() . ':</info>' .
-            ' <comment>seeded' .
-            ' ' . sprintf('%.4fs', $end - $start) . '</comment>'
-        );
-    }
-
-    /**
-     * Rollback an environment to the specified version.
-     *
-     * @param string $environment Environment
-     * @param int|string $target
-     * @param bool $force
-     * @param bool $targetMustMatchVersion
-     * @return void
-     */
-    public function rollback($environment, $target = null, $force = false, $targetMustMatchVersion = true)
-    {
-        // note that the migrations are indexed by name (aka creation time) in ascending order
-        $migrations = $this->getMigrations();
-
-        // note that the version log are also indexed by name with the proper ascending order according to the version order
-        $executedVersions = $this->getEnvironment($environment)->getVersionLog();
-
-        // get a list of migrations sorted in the opposite way of the executed versions
-        $sortedMigrations = [];
-
-        foreach ($executedVersions as $versionCreationTime => &$executedVersion) {
-            // if we have a date (ie. the target must not match a version) and we are sorting by execution time, we
-            // convert the version start time so we can compare directly with the target date
-            if (!$this->getConfig()->isVersionOrderCreationTime() && !$targetMustMatchVersion) {
-                $dateTime = \DateTime::createFromFormat('Y-m-d H:i:s', $executedVersion['start_time']);
-                $executedVersion['start_time'] = $dateTime->format('YmdHis');
-            }
-
-            if (isset($migrations[$versionCreationTime])) {
-                array_unshift($sortedMigrations, $migrations[$versionCreationTime]);
-            } else {
-                // this means the version is missing so we unset it so that we don't consider it when rolling back
-                // migrations (or choosing the last up version as target)
-                unset($executedVersions[$versionCreationTime]);
-            }
-        }
-
-        if ($target === 'all' || $target === '0') {
-            $target = 0;
-        } elseif (!is_numeric($target) && !is_null($target)) { // try to find a target version based on name
-            // search through the migrations using the name
-            $migrationNames = array_map(function ($item) {
-                return $item['migration_name'];
-            }, $executedVersions);
-            $found = array_search($target, $migrationNames);
-
-            // check on was found
-            if ($found !== false) {
-                $target = (string)$found;
-            } else {
-                $this->getOutput()->writeln("<error>No migration found with name ($target)</error>");
-
-                return;
-            }
-        }
-
-        // Check we have at least 1 migration to revert
-        $executedVersionCreationTimes = array_keys($executedVersions);
-        if (empty($executedVersionCreationTimes) || $target == end($executedVersionCreationTimes)) {
-            $this->getOutput()->writeln('<error>No migrations to rollback</error>');
-
-            return;
-        }
-
-        // If no target was supplied, revert the last migration
-        if ($target === null) {
-            // Get the migration before the last run migration
-            $prev = count($executedVersionCreationTimes) - 2;
-            $target = $prev >= 0 ? $executedVersionCreationTimes[$prev] : 0;
-        }
-
-        // If the target must match a version, check the target version exists
-        if ($targetMustMatchVersion && 0 !== $target && !isset($migrations[$target])) {
-            $this->getOutput()->writeln("<error>Target version ($target) not found</error>");
-
-            return;
-        }
-
-        // Rollback all versions until we find the wanted rollback target
-        $rollbacked = false;
-
-        foreach ($sortedMigrations as $migration) {
-            if ($targetMustMatchVersion && $migration->getVersion() == $target) {
-                break;
-            }
-
-            if (in_array($migration->getVersion(), $executedVersionCreationTimes)) {
-                $executedVersion = $executedVersions[$migration->getVersion()];
-
-                if (!$targetMustMatchVersion) {
-                    if (($this->getConfig()->isVersionOrderCreationTime() && $executedVersion['version'] <= $target) ||
-                        (!$this->getConfig()->isVersionOrderCreationTime() && $executedVersion['start_time'] <= $target)) {
-                        break;
-                    }
-                }
-
-                if (0 != $executedVersion['breakpoint'] && !$force) {
-                    $this->getOutput()->writeln('<error>Breakpoint reached. Further rollbacks inhibited.</error>');
-                    break;
-                }
-                $this->executeMigration($environment, $migration, MigrationInterface::DOWN);
-                $rollbacked = true;
-            }
-        }
-
-        if (!$rollbacked) {
-            $this->getOutput()->writeln('<error>No migrations to rollback</error>');
-        }
-    }
-
-    /**
-     * Run database seeders against an environment.
-     *
-     * @param string $environment Environment
-     * @param string $seed Seeder
-     * @return void
-     */
-    public function seed($environment, $seed = null)
-    {
-        $seeds = $this->getSeeds();
-
-        if ($seed === null) {
-            // run all seeders
-            foreach ($seeds as $seeder) {
-                if (array_key_exists($seeder->getName(), $seeds)) {
-                    $this->executeSeed($environment, $seeder);
-                }
-            }
-        } else {
-            // run only one seeder
-            if (array_key_exists($seed, $seeds)) {
-                $this->executeSeed($environment, $seeds[$seed]);
-            } else {
-                throw new \InvalidArgumentException(sprintf('The seed class "%s" does not exist', $seed));
-            }
-        }
     }
 
     /**
@@ -756,152 +568,6 @@ class Manager
     }
 
     /**
-     * Sets the database seeders.
-     *
-     * @param array $seeds Seeders
-     * @return \Phinx\Migration\Manager
-     */
-    public function setSeeds(array $seeds)
-    {
-        $this->seeds = $seeds;
-
-        return $this;
-    }
-
-    /**
-     * Get seed dependencies instances from seed dependency array
-     *
-     * @param AbstractSeed $seed Seed
-     *
-     * @return AbstractSeed[]
-     */
-    private function getSeedDependenciesInstances(AbstractSeed $seed)
-    {
-        $dependenciesInstances = [];
-        $dependencies = $seed->getDependencies();
-        if (!empty($dependencies)) {
-            foreach ($dependencies as $dependency) {
-                foreach ($this->seeds as $seed) {
-                    if (get_class($seed) === $dependency) {
-                        $dependenciesInstances[get_class($seed)] = $seed;
-                    }
-                }
-            }
-        }
-
-        return $dependenciesInstances;
-    }
-
-    /**
-     * Order seeds by dependencies
-     *
-     * @param AbstractSeed[] $seeds Seeds
-     *
-     * @return AbstractSeed[]
-     */
-    private function orderSeedsByDependencies(array $seeds)
-    {
-        $orderedSeeds = [];
-        foreach ($seeds as $seed) {
-            $key = get_class($seed);
-            $dependencies = $this->getSeedDependenciesInstances($seed);
-            if (!empty($dependencies)) {
-                $orderedSeeds[$key] = $seed;
-                $orderedSeeds = array_merge($this->orderSeedsByDependencies($dependencies), $orderedSeeds);
-            } else {
-                $orderedSeeds[$key] = $seed;
-            }
-        }
-
-        return $orderedSeeds;
-    }
-
-    /**
-     * Gets an array of database seeders.
-     *
-     * @throws \InvalidArgumentException
-     * @return \Phinx\Seed\AbstractSeed[]
-     */
-    public function getSeeds()
-    {
-        if ($this->seeds === null) {
-            $phpFiles = $this->getSeedFiles();
-
-            // filter the files to only get the ones that match our naming scheme
-            $fileNames = [];
-            /** @var \Phinx\Seed\AbstractSeed[] $seeds */
-            $seeds = [];
-
-            foreach ($phpFiles as $filePath) {
-                if (Util::isValidSeedFileName(basename($filePath))) {
-                    $config = $this->getConfig();
-                    $namespace = $config instanceof NamespaceAwareInterface ? $config->getSeedNamespaceByPath(dirname($filePath)) : null;
-
-                    // convert the filename to a class name
-                    $class = ($namespace === null ? '' : $namespace . '\\') . pathinfo($filePath, PATHINFO_FILENAME);
-                    $fileNames[$class] = basename($filePath);
-
-                    // load the seed file
-                    /** @noinspection PhpIncludeInspection */
-                    require_once $filePath;
-                    if (!class_exists($class)) {
-                        throw new \InvalidArgumentException(sprintf(
-                            'Could not find class "%s" in file "%s"',
-                            $class,
-                            $filePath
-                        ));
-                    }
-
-                    // instantiate it
-                    $seed = new $class($this->getInput(), $this->getOutput());
-
-                    if (!($seed instanceof AbstractSeed)) {
-                        throw new \InvalidArgumentException(sprintf(
-                            'The class "%s" in file "%s" must extend \Phinx\Seed\AbstractSeed',
-                            $class,
-                            $filePath
-                        ));
-                    }
-
-                    $seeds[$class] = $seed;
-                }
-            }
-
-            ksort($seeds);
-            $this->setSeeds($seeds);
-        }
-
-        $this->seeds = $this->orderSeedsByDependencies($this->seeds);
-        return $this->seeds;
-    }
-
-    /**
-     * Returns a list of seed files found in the provided seed paths.
-     *
-     * @return string[]
-     */
-    protected function getSeedFiles()
-    {
-        $config = $this->getConfig();
-        $paths = $config->getSeedPaths();
-        $files = [];
-
-        foreach ($paths as $path) {
-            $files = array_merge(
-                $files,
-                Util::glob($path . DIRECTORY_SEPARATOR . '*.php')
-            );
-        }
-        // glob() can return the same file multiple times
-        // This will cause the migration to fail with a
-        // false assumption of duplicate migrations
-        // http://php.net/manual/en/function.glob.php#110340
-        $files = array_unique($files);
-
-        return $files;
-    }
-
-    /**
      * Sets the config.
      *
      * @param  \Phinx\Config\ConfigInterface $config Configuration Object
@@ -924,60 +590,4 @@ class Manager
         return $this->config;
     }
 
-    /**
-     * Toggles the breakpoint for a specific version.
-     *
-     * @param string $environment
-     * @param int $version
-     * @return void
-     */
-    public function toggleBreakpoint($environment, $version)
-    {
-        $migrations = $this->getMigrations();
-        $this->getMigrations();
-        $env = $this->getEnvironment($environment);
-        $versions = $env->getVersionLog();
-
-        if (empty($versions) || empty($migrations)) {
-            return;
-        }
-
-        if ($version === null) {
-            $lastVersion = end($versions);
-            $version = $lastVersion['version'];
-        }
-
-        if (0 != $version && !isset($migrations[$version])) {
-            $this->output->writeln(sprintf(
-                '<comment>warning</comment> %s is not a valid version',
-                $version
-            ));
-
-            return;
-        }
-
-        $env->getAdapter()->toggleBreakpoint($migrations[$version]);
-
-        $versions = $env->getVersionLog();
-
-        $this->getOutput()->writeln(
-            ' Breakpoint ' . ($versions[$version]['breakpoint'] ? 'set' : 'cleared') .
-            ' for <info>' . $version . '</info>' .
-            ' <comment>' . $migrations[$version]->getName() . '</comment>'
-        );
-    }
-
-    /**
-     * Remove all breakpoints
-     *
-     * @param string $environment
-     * @return void
-     */
-    public function removeBreakpoints($environment)
-    {
-        $this->getOutput()->writeln(sprintf(
-            ' %d breakpoints cleared.',
-            $this->getEnvironment($environment)->getAdapter()->resetAllBreakpoints()
-        ));
-    }
 }
