@@ -12,12 +12,11 @@ use Phinx\Config\Config;
  */
 class ConfigMigrationPathsTest extends AbstractConfigTest
 {
-    /**
-     * @expectedException \UnexpectedValueException
-     */
+
     public function testGetMigrationPathsThrowsExceptionForNoPath()
     {
         $config = new Config([]);
+        $this->expectException(\UnexpectedValueException::class);
         $config->getMigrationPaths();
     }
 
@@ -41,7 +40,7 @@ class ConfigMigrationPathsTest extends AbstractConfigTest
         $config = new Config($values);
         $paths = $config->getMigrationPaths();
 
-        $this->assertInternalType('array', $paths);
+        $this->assertIsArray($paths);
         $this->assertCount(1, $paths);
     }
 }
