@@ -32,7 +32,7 @@ class MigrateTest extends TestCase
      */
     protected $output;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->config = new Config([
             'paths' => [
@@ -78,7 +78,7 @@ class MigrateTest extends TestCase
         $commandTester = new CommandTester($command);
         $exitCode = $commandTester->execute(['command' => $command->getName()], ['decorated' => false]);
 
-        $this->assertRegExp('/no environment specified/', $commandTester->getDisplay());
+        $this->assertMatchesRegularExpression('/no environment specified/', $commandTester->getDisplay());
         $this->assertSame(0, $exitCode);
     }
 
@@ -104,7 +104,7 @@ class MigrateTest extends TestCase
         $commandTester = new CommandTester($command);
         $exitCode = $commandTester->execute(['command' => $command->getName(), '--environment' => 'fakeenv'], ['decorated' => false]);
 
-        $this->assertRegExp('/using environment fakeenv/', $commandTester->getDisplay());
+        $this->assertMatchesRegularExpression('/using environment fakeenv/', $commandTester->getDisplay());
         $this->assertSame(1, $exitCode);
     }
 
@@ -130,7 +130,7 @@ class MigrateTest extends TestCase
         $commandTester = new CommandTester($command);
         $exitCode = $commandTester->execute(['command' => $command->getName()], ['decorated' => false]);
 
-        $this->assertRegExp('/using database development/', $commandTester->getDisplay());
+        $this->assertMatchesRegularExpression('/using database development/', $commandTester->getDisplay());
         $this->assertSame(0, $exitCode);
     }
 }
