@@ -3,6 +3,7 @@
 namespace Test\Phinx\Db\Adapter;
 
 use Phinx\Db\Adapter\AdapterFactory;
+use Phinx\Db\Adapter\TimedOutputAdapter;
 use PHPUnit\Framework\TestCase;
 
 class AdapterFactoryTest extends TestCase
@@ -12,12 +13,12 @@ class AdapterFactoryTest extends TestCase
      */
     private $factory;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->factory = AdapterFactory::instance();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         unset($this->factory);
     }
@@ -40,13 +41,11 @@ class AdapterFactoryTest extends TestCase
         $this->assertEquals($adapter, $method->invoke($this->factory, 'test'));
     }
 
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Adapter class "Test\Phinx\Db\Adapter\AdapterFactoryTest" must implement Phinx\Db\Adapter\AdapterInterface
-     */
     public function testRegisterAdapterFailure()
     {
         $adapter = get_class($this);
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Adapter class "Test\Phinx\Db\Adapter\AdapterFactoryTest" must implement Phinx\Db\Adapter\AdapterInterface');
         $this->factory->registerAdapter('test', $adapter);
     }
 
@@ -57,12 +56,10 @@ class AdapterFactoryTest extends TestCase
         $this->assertInstanceOf('Phinx\Db\Adapter\MysqlAdapter', $adapter);
     }
 
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Adapter "bad" has not been registered
-     */
     public function testGetAdapterFailure()
     {
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Adapter "bad" has not been registered');
         $this->factory->getAdapter('bad', []);
     }
 
@@ -73,19 +70,17 @@ class AdapterFactoryTest extends TestCase
         $method = new \ReflectionMethod(get_class($this->factory), 'getWrapperClass');
         $method->setAccessible(true);
 
-        $wrapper = $method->invoke($this->factory, 'proxy');
+        $wrapper = $method->invoke($this->factory, 'timed');
         $this->factory->registerWrapper('test', $wrapper);
 
         $this->assertEquals($wrapper, $method->invoke($this->factory, 'test'));
     }
 
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Wrapper class "Test\Phinx\Db\Adapter\AdapterFactoryTest" must be implement Phinx\Db\Adapter\WrapperInterface
-     */
     public function testRegisterWrapperFailure()
     {
         $wrapper = get_class($this);
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Wrapper class "Test\Phinx\Db\Adapter\AdapterFactoryTest" must be implement Phinx\Db\Adapter\WrapperInterface');
         $this->factory->registerWrapper('test', $wrapper);
     }
 
@@ -96,17 +91,15 @@ class AdapterFactoryTest extends TestCase
 
     public function testGetWrapper()
     {
-        $wrapper = $this->factory->getWrapper('prefix', $this->getAdapterMock());
+        $wrapper = $this->factory->getWrapper('timed', $this->getAdapterMock());
 
-        $this->assertInstanceOf('Phinx\Db\Adapter\TablePrefixAdapter', $wrapper);
+        $this->assertInstanceOf(TimedOutputAdapter::class, $wrapper);
     }
 
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Wrapper "nope" has not been registered
-     */
     public function testGetWrapperFailure()
     {
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Wrapper "nope" has not been registered');
         $this->factory->getWrapper('nope', $this->getAdapterMock());
     }
 }
