@@ -2,47 +2,81 @@
 
 namespace Test\Phinx\Db\Adapter;
 
-use Phinx\Db\Adapter\AdapterInterface;
 use Phinx\Db\Adapter\MysqlAdapter;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Input\InputDefinition;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Console\Output\NullOutput;
-use Symfony\Component\Console\Output\StreamOutput;
+use const MYSQL_DB_CONFIG;
 
 class MysqlAdapterTest extends TestCase
 {
+    private const EXPECTED_SCHEMA_TABLE_STRUCTURE = [
+        [
+            'Field'   => 'version',
+            'Type'    => 'bigint(20)',
+            'Null'    => 'NO',
+            'Key'     => 'PRI',
+            'Default' => NULL,
+            'Extra'   => '',
+        ],
+        [
+            'Field'   => 'migration_name',
+            'Type'    => 'varchar(100)',
+            'Null'    => 'YES',
+            'Key'     => '',
+            'Default' => NULL,
+            'Extra'   => '',
+        ],
+        [
+            'Field'   => 'start_time',
+            'Type'    => 'timestamp',
+            'Null'    => 'YES',
+            'Key'     => '',
+            'Default' => NULL,
+            'Extra'   => '',
+        ],
+
+        [
+            'Field'   => 'end_time',
+            'Type'    => 'timestamp',
+            'Null'    => 'YES',
+            'Key'     => '',
+            'Default' => NULL,
+            'Extra'   => '',
+        ],
+        [
+            'Field'   => 'breakpoint',
+            'Type'    => 'tinyint(1)',
+            'Null'    => 'NO',
+            'Key'     => '',
+            'Default' => '0',
+            'Extra'   => '',
+        ],
+    ];
+
     /**
      * @var \Phinx\Db\Adapter\MysqlAdapter
      */
     private $adapter;
 
-    public function setUp()
+    public function setUp(): void
     {
-        if (!TESTS_PHINX_DB_ADAPTER_MYSQL_ENABLED) {
-            $this->markTestSkipped('Mysql tests disabled. See TESTS_PHINX_DB_ADAPTER_MYSQL_ENABLED constant.');
+        if (!defined('MYSQL_DB_CONFIG')) {
+            $this->markTestSkipped('Mysql tests disabled.');
         }
 
-        $options = [
-            'host' => TESTS_PHINX_DB_ADAPTER_MYSQL_HOST,
-            'name' => TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE,
-            'user' => TESTS_PHINX_DB_ADAPTER_MYSQL_USERNAME,
-            'pass' => TESTS_PHINX_DB_ADAPTER_MYSQL_PASSWORD,
-            'port' => TESTS_PHINX_DB_ADAPTER_MYSQL_PORT
-        ];
-        $this->adapter = new MysqlAdapter($options, new ArrayInput([]), new NullOutput());
+        $this->adapter = new MysqlAdapter(MYSQL_DB_CONFIG, new ArrayInput([]), new NullOutput());
 
         // ensure the database is empty for each test
-        $this->adapter->dropDatabase($options['name']);
-        $this->adapter->createDatabase($options['name']);
+        $database_name = MYSQL_DB_CONFIG['name'];
+        $this->adapter->execute("DROP DATABASE IF EXISTS `$database_name`;");
+        $this->adapter->execute("CREATE DATABASE `$database_name`;");
 
         // leave the adapter in a disconnected state for each test
         $this->adapter->disconnect();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         unset($this->adapter);
     }
@@ -54,6 +88,7 @@ class MysqlAdapterTest extends TestCase
 
     public function testConnectionWithoutPort()
     {
+        $this->markTestSkipped('MySQL is on a non-standard port. Testing without port is going to result in failure');
         $options = $this->adapter->getOptions();
         unset($options['port']);
         $this->adapter->setOptions($options);
@@ -63,9 +98,9 @@ class MysqlAdapterTest extends TestCase
     public function testConnectionWithInvalidCredentials()
     {
         $options = [
-            'host' => TESTS_PHINX_DB_ADAPTER_MYSQL_HOST,
-            'name' => TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE,
-            'port' => TESTS_PHINX_DB_ADAPTER_MYSQL_PORT,
+            'host' => MYSQL_DB_CONFIG['host'],
+            'name' => MYSQL_DB_CONFIG['name'],
+            'port' => MYSQL_DB_CONFIG['port'],
             'user' => 'invaliduser',
             'pass' => 'invalidpass'
         ];
@@ -80,21 +115,19 @@ class MysqlAdapterTest extends TestCase
                 $e,
                 'Expected exception of type InvalidArgumentException, got ' . get_class($e)
             );
-            $this->assertRegExp('/There was a problem connecting to the database/', $e->getMessage());
+            $this->assertMatchesRegularExpression('/There was a problem connecting to the database/', $e->getMessage());
         }
     }
 
     public function testConnectionWithSocketConnection()
     {
-        if (!TESTS_PHINX_DB_ADAPTER_MYSQL_UNIX_SOCKET) {
-            $this->markTestSkipped('MySQL socket connection skipped. See TESTS_PHINX_DB_ADAPTER_MYSQL_UNIX_SOCKET constant.');
-        }
+        $this->markTestSkipped('MySQL socket connection skipped. We don\'t support UNIX sockets.');
 
         $options = [
-            'name' => TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE,
-            'user' => TESTS_PHINX_DB_ADAPTER_MYSQL_USERNAME,
-            'pass' => TESTS_PHINX_DB_ADAPTER_MYSQL_PASSWORD,
-            'unix_socket' => TESTS_PHINX_DB_ADAPTER_MYSQL_UNIX_SOCKET,
+            'name' => MYSQL_DB_CONFIG['name'],
+            'user' => MYSQL_DB_CONFIG['user'],
+            'pass' => MYSQL_DB_CONFIG['pass'],
+            'unix_socket' => '',
         ];
 
         $adapter = new MysqlAdapter($options, new ArrayInput([]), new NullOutput());
@@ -105,20 +138,69 @@ class MysqlAdapterTest extends TestCase
 
     public function testCreatingTheSchemaTableOnConnect()
     {
+        $schema_table_name = $this->adapter->getSchemaTableName();
         $this->adapter->connect();
         $this->assertTrue($this->adapter->hasTable($this->adapter->getSchemaTableName()));
-        $this->adapter->dropTable($this->adapter->getSchemaTableName());
+        $this->adapter->execute("DROP TABLE `$schema_table_name`;");
         $this->assertFalse($this->adapter->hasTable($this->adapter->getSchemaTableName()));
         $this->adapter->disconnect();
         $this->adapter->connect();
         $this->assertTrue($this->adapter->hasTable($this->adapter->getSchemaTableName()));
     }
 
-    public function testSchemaTableIsCreatedWithPrimaryKey()
+    public function testSchemaTableIsCreatedWithExpectedStructureIfMissing()
     {
+        $this->adapter->setSchemaTableName('db_migrations');
         $this->adapter->connect();
-        $table = new \Phinx\Db\Table($this->adapter->getSchemaTableName(), [], $this->adapter);
-        $this->assertTrue($this->adapter->hasIndex($this->adapter->getSchemaTableName(), ['version']));
+        $schema = $this->adapter->query('DESCRIBE `db_migrations`')->fetchAll(\PDO::FETCH_ASSOC);
+        $this->assertSame(self::EXPECTED_SCHEMA_TABLE_STRUCTURE, $schema);
+    }
+
+    public function providerOldSchemaTablesToUpgrade()
+    {
+        return [
+            'no breakpoint'     => [
+                <<<SQL
+                CREATE TABLE `old_migrations_table` (
+                    `version` bigint(20) NOT NULL,
+                    `migration_name` varchar(100) DEFAULT NULL,
+                    `start_time` timestamp NULL DEFAULT NULL,
+                    `end_time` timestamp NULL DEFAULT NULL,
+                    PRIMARY KEY (`version`)
+                ) ENGINE=InnoDB DEFAULT CHARSET=utf8
+                SQL
+    ,
+            ],
+            'no migration_name' => [
+                <<<SQL
+                CREATE TABLE `old_migrations_table` (
+                    `version` bigint(20) NOT NULL,
+                    `start_time` timestamp NULL DEFAULT NULL,
+                    `end_time` timestamp NULL DEFAULT NULL,
+                    `breakpoint` tinyint(1) NOT NULL DEFAULT '0',
+                    PRIMARY KEY (`version`)
+                ) ENGINE=InnoDB DEFAULT CHARSET=utf8
+                SQL
+    ,
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider providerOldSchemaTablesToUpgrade
+     */
+    public function testConnectUpgradesSchemaTableIfRequired($old_schema_create_statement)
+    {
+        // Create a table with the old schema
+        $this->adapter->connect();
+        $this->adapter->execute($old_schema_create_statement);
+        $this->adapter->disconnect();
+
+        $this->adapter->setSchemaTableName('old_migrations_table');
+        $this->assertSame(
+            self::EXPECTED_SCHEMA_TABLE_STRUCTURE,
+            $this->adapter->query('DESCRIBE `old_migrations_table`')->fetchAll(\PDO::FETCH_ASSOC)
+        );
     }
 
     public function testQuoteTableName()
@@ -131,1002 +213,16 @@ class MysqlAdapterTest extends TestCase
         $this->assertEquals('`test_column`', $this->adapter->quoteColumnName('test_column'));
     }
 
-    public function testCreateTable()
-    {
-        $table = new \Phinx\Db\Table('ntable', [], $this->adapter);
-        $table->addColumn('realname', 'string')
-              ->addColumn('email', 'integer')
-              ->save();
-        $this->assertTrue($this->adapter->hasTable('ntable'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'id'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'realname'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'email'));
-        $this->assertFalse($this->adapter->hasColumn('ntable', 'address'));
-    }
-
-    public function testCreateTableWithComment()
-    {
-        $tableComment = 'Table comment';
-        $table = new \Phinx\Db\Table('ntable', ['comment' => $tableComment], $this->adapter);
-        $table->addColumn('realname', 'string')
-              ->save();
-        $this->assertTrue($this->adapter->hasTable('ntable'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'id'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'realname'));
-        $this->assertFalse($this->adapter->hasColumn('ntable', 'address'));
-
-        $rows = $this->adapter->fetchAll(sprintf(
-            "SELECT table_comment FROM INFORMATION_SCHEMA.TABLES WHERE table_schema='%s' AND table_name='ntable'",
-            TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE
-        ));
-        $comment = $rows[0];
-
-        $this->assertEquals($tableComment, $comment['table_comment'], 'Dont set table comment correctly');
-    }
-
-    public function testCreateTableWithForeignKeys()
-    {
-        $tag_table = new \Phinx\Db\Table('ntable_tag', [], $this->adapter);
-        $tag_table->addColumn('realname', 'string')
-                  ->save();
-
-        $table = new \Phinx\Db\Table('ntable', [], $this->adapter);
-        $table->addColumn('realname', 'string')
-              ->addColumn('tag_id', 'integer')
-              ->addForeignKey('tag_id', 'ntable_tag', 'id', ['delete' => 'NO_ACTION', 'update' => 'NO_ACTION'])
-              ->save();
-
-        $this->assertTrue($this->adapter->hasTable('ntable'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'id'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'realname'));
-        $this->assertFalse($this->adapter->hasColumn('ntable', 'address'));
-
-        $rows = $this->adapter->fetchAll(sprintf(
-            "SELECT table_name, column_name, referenced_table_name, referenced_column_name
-             FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
-             WHERE table_schema='%s' AND REFERENCED_TABLE_NAME='ntable_tag'",
-            TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE
-        ));
-        $foreignKey = $rows[0];
-
-        $this->assertEquals($foreignKey['table_name'], 'ntable');
-        $this->assertEquals($foreignKey['column_name'], 'tag_id');
-        $this->assertEquals($foreignKey['referenced_table_name'], 'ntable_tag');
-        $this->assertEquals($foreignKey['referenced_column_name'], 'id');
-    }
-
-    public function testCreateTableCustomIdColumn()
-    {
-        $table = new \Phinx\Db\Table('ntable', ['id' => 'custom_id'], $this->adapter);
-        $table->addColumn('realname', 'string')
-              ->addColumn('email', 'integer')
-              ->save();
-        $this->assertTrue($this->adapter->hasTable('ntable'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'custom_id'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'realname'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'email'));
-        $this->assertFalse($this->adapter->hasColumn('ntable', 'address'));
-    }
-
-    public function testCreateTableWithNoOptions()
-    {
-        $this->markTestIncomplete();
-        //$this->adapter->createTable('ntable', )
-    }
-
-    public function testCreateTableWithNoPrimaryKey()
-    {
-        $options = [
-            'id' => false
-        ];
-        $table = new \Phinx\Db\Table('atable', $options, $this->adapter);
-        $table->addColumn('user_id', 'integer')
-              ->save();
-        $this->assertFalse($this->adapter->hasColumn('atable', 'id'));
-    }
-
-    public function testCreateTableWithMultiplePrimaryKeys()
-    {
-        $options = [
-            'id' => false,
-            'primary_key' => ['user_id', 'tag_id']
-        ];
-        $table = new \Phinx\Db\Table('table1', $options, $this->adapter);
-        $table->addColumn('user_id', 'integer')
-              ->addColumn('tag_id', 'integer')
-              ->save();
-        $this->assertTrue($this->adapter->hasIndex('table1', ['user_id', 'tag_id']));
-        $this->assertTrue($this->adapter->hasIndex('table1', ['USER_ID', 'tag_id']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['tag_id', 'user_id']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['tag_id', 'user_email']));
-    }
-
-    public function testCreateTableWithMultipleIndexes()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->addColumn('name', 'string')
-              ->addIndex('email')
-              ->addIndex('name')
-              ->save();
-        $this->assertTrue($this->adapter->hasIndex('table1', ['email']));
-        $this->assertTrue($this->adapter->hasIndex('table1', ['name']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['email', 'user_email']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['email', 'user_name']));
-    }
-
-    public function testCreateTableWithUniqueIndexes()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->addIndex('email', ['unique' => true])
-              ->save();
-        $this->assertTrue($this->adapter->hasIndex('table1', ['email']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['email', 'user_email']));
-    }
-
-    public function testCreateTableWithFullTextIndex()
-    {
-        $table = new \Phinx\Db\Table('table1', ['engine' => 'MyISAM'], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->addIndex('email', ['type' => 'fulltext'])
-              ->save();
-        $this->assertTrue($this->adapter->hasIndex('table1', ['email']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['email', 'user_email']));
-    }
-
-    public function testCreateTableWithNamedIndex()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->addIndex('email', ['name' => 'myemailindex'])
-              ->save();
-        $this->assertTrue($this->adapter->hasIndex('table1', ['email']));
-        $this->assertFalse($this->adapter->hasIndex('table1', ['email', 'user_email']));
-        $this->assertTrue($this->adapter->hasIndexByName('table1', 'myemailindex'));
-    }
-
-    public function testCreateTableWithMultiplePKsAndUniqueIndexes()
-    {
-        $this->markTestIncomplete();
-    }
-
-    public function testCreateTableWithMyISAMEngine()
-    {
-        $table = new \Phinx\Db\Table('ntable', ['engine' => 'MyISAM'], $this->adapter);
-        $table->addColumn('realname', 'string')
-              ->save();
-        $this->assertTrue($this->adapter->hasTable('ntable'));
-        $row = $this->adapter->fetchRow(sprintf("SHOW TABLE STATUS WHERE Name = '%s'", 'ntable'));
-        $this->assertEquals('MyISAM', $row['Engine']);
-    }
-
-    public function testCreateTableAndInheritDefaultCollation()
-    {
-        $options = [
-            'host' => TESTS_PHINX_DB_ADAPTER_MYSQL_HOST,
-            'name' => TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE,
-            'user' => TESTS_PHINX_DB_ADAPTER_MYSQL_USERNAME,
-            'pass' => TESTS_PHINX_DB_ADAPTER_MYSQL_PASSWORD,
-            'port' => TESTS_PHINX_DB_ADAPTER_MYSQL_PORT,
-            'charset' => 'utf8',
-            'collation' => 'utf8_unicode_ci',
-        ];
-        $adapter = new MysqlAdapter($options, new ArrayInput([]), new NullOutput());
-
-        // Ensure the database is empty and the adapter is in a disconnected state
-        $adapter->dropDatabase($options['name']);
-        $adapter->createDatabase($options['name']);
-        $adapter->disconnect();
-
-        $table = new \Phinx\Db\Table('table_with_default_collation', [], $adapter);
-        $table->addColumn('name', 'string')
-              ->save();
-        $this->assertTrue($adapter->hasTable('table_with_default_collation'));
-        $row = $adapter->fetchRow(sprintf("SHOW TABLE STATUS WHERE Name = '%s'", 'table_with_default_collation'));
-        $this->assertEquals('utf8_unicode_ci', $row['Collation']);
-    }
-
-    public function testCreateTableWithLatin1Collate()
-    {
-        $table = new \Phinx\Db\Table('latin1_table', ['collation' => 'latin1_general_ci'], $this->adapter);
-        $table->addColumn('name', 'string')
-              ->save();
-        $this->assertTrue($this->adapter->hasTable('latin1_table'));
-        $row = $this->adapter->fetchRow(sprintf("SHOW TABLE STATUS WHERE Name = '%s'", 'latin1_table'));
-        $this->assertEquals('latin1_general_ci', $row['Collation']);
-    }
-
-    public function testCreateTableWithUnsignedPK()
-    {
-        $table = new \Phinx\Db\Table('ntable', ['signed' => false], $this->adapter);
-        $table->addColumn('realname', 'string')
-              ->addColumn('email', 'integer')
-              ->save();
-        $this->assertTrue($this->adapter->hasTable('ntable'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'id'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'realname'));
-        $this->assertTrue($this->adapter->hasColumn('ntable', 'email'));
-        $this->assertFalse($this->adapter->hasColumn('ntable', 'address'));
-    }
-
-    public function testRenameTable()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertTrue($this->adapter->hasTable('table1'));
-        $this->assertFalse($this->adapter->hasTable('table2'));
-        $this->adapter->renameTable('table1', 'table2');
-        $this->assertFalse($this->adapter->hasTable('table1'));
-        $this->assertTrue($this->adapter->hasTable('table2'));
-    }
-
-    public function testAddColumn()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('email'));
-        $table->addColumn('email', 'string')
-              ->save();
-        $this->assertTrue($table->hasColumn('email'));
-        $table->addColumn('realname', 'string', ['after' => 'id'])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('realname', $rows[1]['Field']);
-    }
-
-    public function testAddColumnWithDefaultValue()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $table->addColumn('default_zero', 'string', ['default' => 'test'])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals("test", $rows[1]['Default']);
-    }
-
-    public function testAddColumnWithDefaultZero()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $table->addColumn('default_zero', 'integer', ['default' => 0])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertNotNull($rows[1]['Default']);
-        $this->assertEquals("0", $rows[1]['Default']);
-    }
-
-    public function testAddColumnWithDefaultEmptyString()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $table->addColumn('default_empty', 'string', ['default' => ''])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('', $rows[1]['Default']);
-    }
-
-    public function testAddColumnWithDefaultBoolean()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $table->addColumn('default_true', 'boolean', ['default' => true])
-              ->addColumn('default_false', 'boolean', ['default' => false])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('1', $rows[1]['Default']);
-        $this->assertEquals('0', $rows[2]['Default']);
-    }
-
-    public function testAddIntegerColumnWithDefaultSigned()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('user_id'));
-        $table->addColumn('user_id', 'integer')
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('int(11)', $rows[1]['Type']);
-    }
-
-    public function testAddIntegerColumnWithSignedEqualsFalse()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('user_id'));
-        $table->addColumn('user_id', 'integer', ['signed' => false])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('int(11) unsigned', $rows[1]['Type']);
-    }
-
-    public function testAddBooleanColumnWithSignedEqualsFalse()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('test_boolean'));
-        $table->addColumn('test_boolean', 'boolean', ['signed' => false])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('tinyint(1) unsigned', $rows[1]['Type']);
-    }
-
-    public function testAddStringColumnWithSignedEqualsFalse()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('user_id'));
-        $table->addColumn('user_id', 'string', ['signed' => false])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('varchar(255)', $rows[1]['Type']);
-    }
-
-    public function testAddStringColumnWithCustomCollation()
-    {
-        $table = new \Phinx\Db\Table('table_custom_collation', ['collation' => 'utf8_general_ci'], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('string_collation_default'));
-        $this->assertFalse($table->hasColumn('string_collation_custom'));
-        $table->addColumn('string_collation_default', 'string', [])->save();
-        $table->addColumn('string_collation_custom', 'string', ['collation' => 'utf8mb4_unicode_ci'])->save();
-        $rows = $this->adapter->fetchAll('SHOW FULL COLUMNS FROM table_custom_collation');
-        $this->assertEquals('utf8_general_ci', $rows[1]['Collation']);
-        $this->assertEquals('utf8mb4_unicode_ci', $rows[2]['Collation']);
-    }
-
-    public function testRenameColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->save();
-        $this->assertTrue($this->adapter->hasColumn('t', 'column1'));
-        $this->assertFalse($this->adapter->hasColumn('t', 'column2'));
-        $this->adapter->renameColumn('t', 'column1', 'column2');
-        $this->assertFalse($this->adapter->hasColumn('t', 'column1'));
-        $this->assertTrue($this->adapter->hasColumn('t', 'column2'));
-    }
-
-    public function testRenamingANonExistentColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->save();
-
-        try {
-            $this->adapter->renameColumn('t', 'column2', 'column1');
-            $this->fail('Expected the adapter to throw an exception');
-        } catch (\InvalidArgumentException $e) {
-            $this->assertInstanceOf(
-                'InvalidArgumentException',
-                $e,
-                'Expected exception of type InvalidArgumentException, got ' . get_class($e)
-            );
-            $this->assertEquals('The specified column doesn\'t exist: column2', $e->getMessage());
-        }
-    }
-
-    public function testChangeColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->save();
-        $this->assertTrue($this->adapter->hasColumn('t', 'column1'));
-        $newColumn1 = new \Phinx\Db\Table\Column();
-        $newColumn1->setType('string');
-        $table->changeColumn('column1', $newColumn1);
-        $this->assertTrue($this->adapter->hasColumn('t', 'column1'));
-        $newColumn2 = new \Phinx\Db\Table\Column();
-        $newColumn2->setName('column2')
-                   ->setType('string');
-        $table->changeColumn('column1', $newColumn2);
-        $this->assertFalse($this->adapter->hasColumn('t', 'column1'));
-        $this->assertTrue($this->adapter->hasColumn('t', 'column2'));
-    }
-
-    public function testChangeColumnDefaultValue()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string', ['default' => 'test'])
-              ->save();
-        $newColumn1 = new \Phinx\Db\Table\Column();
-        $newColumn1->setDefault('test1')
-                   ->setType('string');
-        $table->changeColumn('column1', $newColumn1);
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM t');
-        $this->assertNotNull($rows[1]['Default']);
-        $this->assertEquals("test1", $rows[1]['Default']);
-    }
-
-    public function testChangeColumnDefaultToZero()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'integer')
-              ->save();
-        $newColumn1 = new \Phinx\Db\Table\Column();
-        $newColumn1->setDefault(0)
-                   ->setType('integer');
-        $table->changeColumn('column1', $newColumn1);
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM t');
-        $this->assertNotNull($rows[1]['Default']);
-        $this->assertEquals("0", $rows[1]['Default']);
-    }
-
-    public function testChangeColumnDefaultToNull()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string', ['default' => 'test'])
-              ->save();
-        $newColumn1 = new \Phinx\Db\Table\Column();
-        $newColumn1->setDefault(null)
-                   ->setType('string');
-        $table->changeColumn('column1', $newColumn1);
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM t');
-        $this->assertNull($rows[1]['Default']);
-    }
-
-    public function testLongTextColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'text', ['limit' => MysqlAdapter::TEXT_LONG])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('longtext', $sqlType['name']);
-    }
-
-    public function testMediumTextColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'text', ['limit' => MysqlAdapter::TEXT_MEDIUM])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('mediumtext', $sqlType['name']);
-    }
-
-    public function testTinyTextColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'text', ['limit' => MysqlAdapter::TEXT_TINY])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('tinytext', $sqlType['name']);
-    }
-
-    public function testBigIntegerColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'integer', ['limit' => MysqlAdapter::INT_BIG])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('bigint', $sqlType['name']);
-    }
-
-    public function testMediumIntegerColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'integer', ['limit' => MysqlAdapter::INT_MEDIUM])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('mediumint', $sqlType['name']);
-    }
-
-    public function testSmallIntegerColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'integer', ['limit' => MysqlAdapter::INT_SMALL])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('smallint', $sqlType['name']);
-    }
-
-    public function testTinyIntegerColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'integer', ['limit' => MysqlAdapter::INT_TINY])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals('tinyint', $sqlType['name']);
-    }
-
-    public function testIntegerColumnLimit()
-    {
-        $limit = 8;
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'integer', ['limit' => $limit])
-              ->save();
-        $columns = $table->getColumns();
-        $sqlType = $this->adapter->getSqlType($columns[1]->getType(), $columns[1]->getLimit());
-        $this->assertEquals($limit, $sqlType['limit']);
-    }
-
-    public function testDropColumn()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->save();
-        $this->assertTrue($this->adapter->hasColumn('t', 'column1'));
-        $this->adapter->dropColumn('t', 'column1');
-        $this->assertFalse($this->adapter->hasColumn('t', 'column1'));
-    }
-
-    public function testGetColumns()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->addColumn('column2', 'integer')
-              ->addColumn('column3', 'biginteger')
-              ->addColumn('column4', 'text')
-              ->addColumn('column5', 'float')
-              ->addColumn('column6', 'decimal')
-              ->addColumn('column7', 'datetime')
-              ->addColumn('column8', 'time')
-              ->addColumn('column9', 'timestamp')
-              ->addColumn('column10', 'date')
-              ->addColumn('column11', 'binary')
-              ->addColumn('column12', 'boolean')
-              ->addColumn('column13', 'string', ['limit' => 10])
-              ->addColumn('column15', 'integer', ['limit' => 10])
-              ->addColumn('column16', 'geometry')
-              ->addColumn('column17', 'point')
-              ->addColumn('column18', 'linestring')
-              ->addColumn('column19', 'polygon')
-              ->addColumn('column20', 'uuid')
-              ->addColumn('column21', 'set', ['values' => "one, two"])
-              ->addColumn('column22', 'enum', ['values' => ['three', 'four']]);
-        $pendingColumns = $table->getPendingColumns();
-        $table->save();
-        $columns = $this->adapter->getColumns('t');
-        $this->assertCount(count($pendingColumns) + 1, $columns);
-        for ($i = 0; $i++; $i < count($pendingColumns)) {
-            $this->assertEquals($pendingColumns[$i], $columns[$i + 1]);
-        }
-    }
-
-    public function testDescribeTable()
-    {
-        $table = new \Phinx\Db\Table('t', [], $this->adapter);
-        $table->addColumn('column1', 'string');
-        $table->save();
-
-        $described = $this->adapter->describeTable('t');
-
-        $this->assertContains($described['TABLE_TYPE'], ['VIEW', 'BASE TABLE']);
-        $this->assertEquals($described['TABLE_NAME'], 't');
-        $this->assertEquals($described['TABLE_SCHEMA'], TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE);
-        $this->assertEquals($described['TABLE_ROWS'], 0);
-    }
-
-    public function testGetColumnsReservedTableName()
-    {
-        $table = new \Phinx\Db\Table('group', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->addColumn('column2', 'integer')
-              ->addColumn('column3', 'biginteger')
-              ->addColumn('column4', 'text')
-              ->addColumn('column5', 'float')
-              ->addColumn('column6', 'decimal')
-              ->addColumn('column7', 'datetime')
-              ->addColumn('column8', 'time')
-              ->addColumn('column9', 'timestamp')
-              ->addColumn('column10', 'date')
-              ->addColumn('column11', 'binary')
-              ->addColumn('column12', 'boolean')
-              ->addColumn('column13', 'string', ['limit' => 10])
-              ->addColumn('column15', 'integer', ['limit' => 10])
-              ->addColumn('column16', 'geometry')
-              ->addColumn('column17', 'point')
-              ->addColumn('column18', 'linestring')
-              ->addColumn('column19', 'polygon')
-              ->addColumn('column20', 'uuid')
-              ->addColumn('column21', 'set', ['values' => "one, two"])
-              ->addColumn('column22', 'enum', ['values' => ['three', 'four']]);
-        $pendingColumns = $table->getPendingColumns();
-        $table->save();
-        $columns = $this->adapter->getColumns('group');
-        $this->assertCount(count($pendingColumns) + 1, $columns);
-        for ($i = 0; $i++; $i < count($pendingColumns)) {
-            $this->assertEquals($pendingColumns[$i], $columns[$i + 1]);
-        }
-    }
-
-    public function testAddIndex()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->save();
-        $this->assertFalse($table->hasIndex('email'));
-        $table->addIndex('email')
-              ->save();
-        $this->assertTrue($table->hasIndex('email'));
-    }
-
-    public function testAddIndexWithLimit()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-            ->save();
-        $this->assertFalse($table->hasIndex('email'));
-        $table->addIndex('email', ['limit' => 50])
-            ->save();
-        $this->assertTrue($table->hasIndex('email'));
-        $index_data = $this->adapter->query(sprintf(
-            'SELECT SUB_PART FROM information_schema.STATISTICS WHERE TABLE_SCHEMA = "%s" AND TABLE_NAME = "table1" AND INDEX_NAME = "email"',
-            TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE
-        ))->fetch(\PDO::FETCH_ASSOC);
-        $expected_limit = $index_data['SUB_PART'];
-        $this->assertEquals($expected_limit, 50);
-    }
-
-    public function testDropIndex()
-    {
-        // single column index
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->addIndex('email')
-              ->save();
-        $this->assertTrue($table->hasIndex('email'));
-        $this->adapter->dropIndex($table->getName(), 'email');
-        $this->assertFalse($table->hasIndex('email'));
-
-        // multiple column index
-        $table2 = new \Phinx\Db\Table('table2', [], $this->adapter);
-        $table2->addColumn('fname', 'string')
-               ->addColumn('lname', 'string')
-               ->addIndex(['fname', 'lname'])
-               ->save();
-        $this->assertTrue($table2->hasIndex(['fname', 'lname']));
-        $this->adapter->dropIndex($table2->getName(), ['fname', 'lname']);
-        $this->assertFalse($table2->hasIndex(['fname', 'lname']));
-
-        // index with name specified, but dropping it by column name
-        $table3 = new \Phinx\Db\Table('table3', [], $this->adapter);
-        $table3->addColumn('email', 'string')
-              ->addIndex('email', ['name' => 'someindexname'])
-              ->save();
-        $this->assertTrue($table3->hasIndex('email'));
-        $this->adapter->dropIndex($table3->getName(), 'email');
-        $this->assertFalse($table3->hasIndex('email'));
-
-        // multiple column index with name specified
-        $table4 = new \Phinx\Db\Table('table4', [], $this->adapter);
-        $table4->addColumn('fname', 'string')
-               ->addColumn('lname', 'string')
-               ->addIndex(['fname', 'lname'], ['name' => 'multiname'])
-               ->save();
-        $this->assertTrue($table4->hasIndex(['fname', 'lname']));
-        $this->adapter->dropIndex($table4->getName(), ['fname', 'lname']);
-        $this->assertFalse($table4->hasIndex(['fname', 'lname']));
-
-        // don't drop multiple column index when dropping single column
-        $table2 = new \Phinx\Db\Table('table5', [], $this->adapter);
-        $table2->addColumn('fname', 'string')
-               ->addColumn('lname', 'string')
-               ->addIndex(['fname', 'lname'])
-               ->save();
-        $this->assertTrue($table2->hasIndex(['fname', 'lname']));
-        $this->adapter->dropIndex($table2->getName(), ['fname']);
-        $this->assertTrue($table2->hasIndex(['fname', 'lname']));
-
-        // don't drop multiple column index with name specified when dropping
-        // single column
-        $table4 = new \Phinx\Db\Table('table6', [], $this->adapter);
-        $table4->addColumn('fname', 'string')
-               ->addColumn('lname', 'string')
-               ->addIndex(['fname', 'lname'], ['name' => 'multiname'])
-               ->save();
-        $this->assertTrue($table4->hasIndex(['fname', 'lname']));
-        $this->adapter->dropIndex($table4->getName(), ['fname']);
-        $this->assertTrue($table4->hasIndex(['fname', 'lname']));
-    }
-
-    public function testDropIndexByName()
-    {
-        // single column index
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('email', 'string')
-              ->addIndex('email', ['name' => 'myemailindex'])
-              ->save();
-        $this->assertTrue($table->hasIndex('email'));
-        $this->adapter->dropIndexByName($table->getName(), 'myemailindex');
-        $this->assertFalse($table->hasIndex('email'));
-
-        // multiple column index
-        $table2 = new \Phinx\Db\Table('table2', [], $this->adapter);
-        $table2->addColumn('fname', 'string')
-               ->addColumn('lname', 'string')
-               ->addIndex(['fname', 'lname'], ['name' => 'twocolumnindex'])
-               ->save();
-        $this->assertTrue($table2->hasIndex(['fname', 'lname']));
-        $this->adapter->dropIndexByName($table2->getName(), 'twocolumnindex');
-        $this->assertFalse($table2->hasIndex(['fname', 'lname']));
-    }
-
-    public function testAddForeignKey()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', [], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer')->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-    }
-
-    public function testAddForeignKeyForTableWithUnsignedPK()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', ['signed' => false], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer', ['signed' => false])->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-    }
-
-    public function testDropForeignKey()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', [], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer')->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->adapter->dropForeignKey($table->getName(), ['ref_table_id']);
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-    }
-
-    public function testDropForeignKeyForTableWithUnsignedPK()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', ['signed' => false], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer', ['signed' => false])->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->adapter->dropForeignKey($table->getName(), ['ref_table_id']);
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-    }
-
-    public function testDropForeignKeyAsString()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', [], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer')->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->adapter->dropForeignKey($table->getName(), 'ref_table_id');
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-    }
-
-    public function testHasForeignKey()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', [], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer')->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id2']));
-    }
-
-    public function testHasForeignKeyForTableWithUnsignedPK()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', ['signed' => false], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer', ['signed' => false])->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), ['ref_table_id']));
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id2']));
-    }
-
-    public function testHasForeignKeyAsString()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', [], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer')->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), 'ref_table_id'));
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), 'ref_table_id2'));
-    }
-
-    public function testHasForeignKeyWithConstraint()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', [], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer')->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setConstraint("my_constraint")
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), ['ref_table_id'], 'my_constraint'));
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id'], 'my_constraint2'));
-    }
-
-    public function testHasForeignKeyWithConstraintForTableWithUnsignedPK()
-    {
-        $refTable = new \Phinx\Db\Table('ref_table', ['signed' => false], $this->adapter);
-        $refTable->addColumn('field1', 'string')->save();
-
-        $table = new \Phinx\Db\Table('table', [], $this->adapter);
-        $table->addColumn('ref_table_id', 'integer', ['signed' => false])->save();
-
-        $fk = new \Phinx\Db\Table\ForeignKey();
-        $fk->setReferencedTable($refTable)
-           ->setColumns(['ref_table_id'])
-           ->setConstraint("my_constraint")
-           ->setReferencedColumns(['id']);
-
-        $this->adapter->addForeignKey($table, $fk);
-        $this->assertTrue($this->adapter->hasForeignKey($table->getName(), ['ref_table_id'], 'my_constraint'));
-        $this->assertFalse($this->adapter->hasForeignKey($table->getName(), ['ref_table_id'], 'my_constraint2'));
-    }
-
-    public function testHasDatabase()
-    {
-        $this->assertFalse($this->adapter->hasDatabase('fake_database_name'));
-        $this->assertTrue($this->adapter->hasDatabase(TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE));
-    }
-
-    public function testDropDatabase()
-    {
-        $this->assertFalse($this->adapter->hasDatabase('phinx_temp_database'));
-        $this->adapter->createDatabase('phinx_temp_database');
-        $this->assertTrue($this->adapter->hasDatabase('phinx_temp_database'));
-        $this->adapter->dropDatabase('phinx_temp_database');
-    }
-
-    public function testAddColumnWithComment()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('column1', 'string', ['comment' => $comment = 'Comments from "column1"'])
-              ->save();
-
-        $rows = $this->adapter->fetchAll(sprintf(
-            "SELECT column_name, column_comment FROM information_schema.columns WHERE table_schema='%s' AND table_name='table1'",
-            TESTS_PHINX_DB_ADAPTER_MYSQL_DATABASE
-        ));
-        $columnWithComment = $rows[1];
-
-        $this->assertEquals($comment, $columnWithComment['column_comment'], 'Dont set column comment correctly');
-    }
-
-    public function testAddGeoSpatialColumns()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('geo_geom'));
-        $table->addColumn('geo_geom', 'geometry')
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals('geometry', $rows[1]['Type']);
-    }
-
-    public function testAddSetColumn()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('set_column'));
-        $table->addColumn('set_column', 'set', ['values' => ['one', 'two']])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals("set('one','two')", $rows[1]['Type']);
-    }
-
-    public function testAddEnumColumn()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->save();
-        $this->assertFalse($table->hasColumn('enum_column'));
-        $table->addColumn('enum_column', 'enum', ['values' => ['one', 'two']])
-              ->save();
-        $rows = $this->adapter->fetchAll('SHOW COLUMNS FROM table1');
-        $this->assertEquals("enum('one','two')", $rows[1]['Type']);
-    }
-
-    public function testEnumColumnValuesFilledUpFromSchema()
-    {
-        // Creating column with values
-        (new \Phinx\Db\Table('table1', [], $this->adapter))
-            ->addColumn('enum_column', 'enum', ['values' => ['one', 'two']])
-            ->save();
-
-        // Reading them back
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $columns = $table->getColumns();
-        $enumColumn = end($columns);
-        $this->assertEquals(AdapterInterface::PHINX_TYPE_ENUM, $enumColumn->getType());
-        $this->assertEquals(['one', 'two'], $enumColumn->getValues());
-    }
-
-    public function testHasColumn()
-    {
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-              ->save();
-
-        $this->assertFalse($table->hasColumn('column2'));
-        $this->assertTrue($table->hasColumn('column1'));
-    }
-
-    public function testHasColumnReservedName()
-    {
-        $tableQuoted = new \Phinx\Db\Table('group', [], $this->adapter);
-        $tableQuoted->addColumn('value', 'string')
-                    ->save();
-
-        $this->assertFalse($tableQuoted->hasColumn('column2'));
-        $this->assertTrue($tableQuoted->hasColumn('value'));
-    }
-
     public function testBulkInsertData()
     {
+        $this->adapter->execute(<<<SQL
+            CREATE TABLE `table1`(
+                `column1` VARCHAR(255),
+                `column2` INT,
+                `column3` VARCHAR(255) NOT NULL DEFAULT 'test'
+            );
+            SQL
+        );
         $data = [
             [
                 'column1' => 'value1',
@@ -1141,14 +237,8 @@ class MysqlAdapterTest extends TestCase
                 'column2' => 3,
             ]
         ];
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-            ->addColumn('column2', 'integer')
-            ->addColumn('column3', 'string', ['default' => 'test'])
-            ->insert($data);
-        $this->adapter->createTable($table);
-        $this->adapter->bulkinsert($table, $table->getData());
-        $table->reset();
+
+        $this->adapter->bulkinsert('table1', $data);
 
         $rows = $this->adapter->fetchAll('SELECT * FROM table1');
         $this->assertEquals('value1', $rows[0]['column1']);
@@ -1163,6 +253,15 @@ class MysqlAdapterTest extends TestCase
 
     public function testInsertData()
     {
+        $this->adapter->execute(<<<SQL
+            CREATE TABLE `table1`(
+                `column1` VARCHAR(255),
+                `column2` INT,
+                `column3` VARCHAR(255) NOT NULL DEFAULT 'test'
+            );
+            SQL
+        );
+
         $data = [
             [
                 'column1' => 'value1',
@@ -1178,12 +277,10 @@ class MysqlAdapterTest extends TestCase
                 'column3' => 'foo',
             ]
         ];
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-        $table->addColumn('column1', 'string')
-            ->addColumn('column2', 'integer')
-            ->addColumn('column3', 'string', ['default' => 'test'])
-            ->insert($data)
-            ->save();
+
+        $this->adapter->insert('table1', $data[0]);
+        $this->adapter->insert('table1', $data[1]);
+        $this->adapter->insert('table1', $data[2]);
 
         $rows = $this->adapter->fetchAll('SELECT * FROM table1');
         $this->assertEquals('value1', $rows[0]['column1']);
@@ -1196,25 +293,4 @@ class MysqlAdapterTest extends TestCase
         $this->assertEquals('foo', $rows[2]['column3']);
     }
 
-    public function testDumpCreateTable()
-    {
-        $inputDefinition = new InputDefinition([new InputOption('dry-run')]);
-        $this->adapter->setInput(new ArrayInput(['--dry-run' => true], $inputDefinition));
-
-        $consoleOutput = new BufferedOutput();
-        $this->adapter->setOutput($consoleOutput);
-
-        $table = new \Phinx\Db\Table('table1', [], $this->adapter);
-
-        $table->addColumn('column1', 'string')
-            ->addColumn('column2', 'integer')
-            ->addColumn('column3', 'string', ['default' => 'test'])
-            ->save();
-
-        $expectedOutput = <<<'OUTPUT'
-CREATE TABLE `table1` (`id` INT(11) NOT NULL AUTO_INCREMENT, `column1` VARCHAR(255) NOT NULL, `column2` INT(11) NOT NULL, `column3` VARCHAR(255) NOT NULL DEFAULT 'test', PRIMARY KEY (`id`)) ENGINE = InnoDB CHARACTER SET utf8 COLLATE utf8_general_ci;
-OUTPUT;
-        $actualOutput = $consoleOutput->fetch();
-        $this->assertContains($expectedOutput, $actualOutput, 'Passing the --dry-run option does not dump create table query to the output');
-    }
 }
