@@ -2,8 +2,7 @@
 
 namespace Test\Phinx\Migration;
 
-use Phinx\Db\Adapter\AdapterInterface;
-use Phinx\Db\Table;
+use Phinx\Db\Adapter\MysqlAdapter;
 use PHPUnit\Framework\TestCase;
 
 class AbstractMigrationTest extends TestCase
@@ -13,13 +12,6 @@ class AbstractMigrationTest extends TestCase
         // stub migration
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
         $this->assertNull($migrationStub->up());
-    }
-
-    public function testDown()
-    {
-        // stub migration
-        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
-        $this->assertNull($migrationStub->down());
     }
 
     public function testAdapterMethods()
@@ -47,9 +39,7 @@ class AbstractMigrationTest extends TestCase
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
 
         // stub output
-        $outputStub = $this->getMockBuilder('\Symfony\Component\Console\Output\OutputInterface')
-            ->setConstructorArgs([[]])
-            ->getMock();
+        $outputStub = $this->getMockBuilder('\Symfony\Component\Console\Output\OutputInterface')->getMock();
 
         // test methods
         $this->assertNull($migrationStub->getOutput());
@@ -60,12 +50,10 @@ class AbstractMigrationTest extends TestCase
     public function testGetInputMethodWithInjectedInput()
     {
         // stub input
-        $inputStub = $this->getMockBuilder('\Symfony\Component\Console\Input\InputInterface')
-            ->setConstructorArgs([[]])
-            ->getMock();
+        $inputStub = $this->getMockBuilder('\Symfony\Component\Console\Input\InputInterface')->getMock();
 
         // stub migration
-        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0, $inputStub, null]);
+        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0, $inputStub, NULL]);
 
         // test methods
         $this->assertNotNull($migrationStub->getInput());
@@ -75,12 +63,10 @@ class AbstractMigrationTest extends TestCase
     public function testGetOutputMethodWithInjectedOutput()
     {
         // stub output
-        $outputStub = $this->getMockBuilder('\Symfony\Component\Console\Output\OutputInterface')
-            ->setConstructorArgs([[]])
-            ->getMock();
+        $outputStub = $this->getMockBuilder('\Symfony\Component\Console\Output\OutputInterface')->getMock();
 
         // stub migration
-        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0, null, $outputStub]);
+        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0, NULL, $outputStub]);
 
         // test methods
         $this->assertNotNull($migrationStub->getOutput());
@@ -90,7 +76,7 @@ class AbstractMigrationTest extends TestCase
     public function testGetName()
     {
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
-        $this->assertContains('AbstractMigration', $migrationStub->getName());
+        $this->assertStringContainsString('AbstractMigration', $migrationStub->getName());
     }
 
     public function testVersionMethods()
@@ -111,8 +97,8 @@ class AbstractMigrationTest extends TestCase
             ->setConstructorArgs([[]])
             ->getMock();
         $adapterStub->expects($this->once())
-                    ->method('execute')
-                    ->will($this->returnValue(2));
+            ->method('execute')
+            ->will($this->returnValue(2));
 
         $migrationStub->setAdapter($adapterStub);
         $this->assertEquals(2, $migrationStub->execute('SELECT FOO FROM BAR'));
@@ -123,16 +109,18 @@ class AbstractMigrationTest extends TestCase
         // stub migration
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
 
+        $queryStub = new class extends \PDOStatement {};
+
         // stub adapter
         $adapterStub = $this->getMockBuilder('\Phinx\Db\Adapter\PdoAdapter')
             ->setConstructorArgs([[]])
             ->getMock();
         $adapterStub->expects($this->once())
-                    ->method('query')
-                    ->will($this->returnValue([['0' => 'bar', 'foo' => 'bar']]));
+            ->method('query')
+            ->will($this->returnValue($queryStub));
 
         $migrationStub->setAdapter($adapterStub);
-        $this->assertEquals([['0' => 'bar', 'foo' => 'bar']], $migrationStub->query('SELECT FOO FROM BAR'));
+        $this->assertEquals($queryStub, $migrationStub->query('SELECT FOO FROM BAR'));
     }
 
     public function testFetchRow()
@@ -145,8 +133,8 @@ class AbstractMigrationTest extends TestCase
             ->setConstructorArgs([[]])
             ->getMock();
         $adapterStub->expects($this->once())
-                    ->method('fetchRow')
-                    ->will($this->returnValue(['0' => 'bar', 'foo' => 'bar']));
+            ->method('fetchRow')
+            ->will($this->returnValue(['0' => 'bar', 'foo' => 'bar']));
 
         $migrationStub->setAdapter($adapterStub);
         $this->assertEquals(['0' => 'bar', 'foo' => 'bar'], $migrationStub->fetchRow('SELECT FOO FROM BAR'));
@@ -162,14 +150,37 @@ class AbstractMigrationTest extends TestCase
             ->setConstructorArgs([[]])
             ->getMock();
         $adapterStub->expects($this->once())
-                    ->method('fetchAll')
-                    ->will($this->returnValue([['0' => 'bar', 'foo' => 'bar']]));
+            ->method('fetchAll')
+            ->will($this->returnValue([['0' => 'bar', 'foo' => 'bar']]));
 
         $migrationStub->setAdapter($adapterStub);
         $this->assertEquals([['0' => 'bar', 'foo' => 'bar']], $migrationStub->fetchAll('SELECT FOO FROM BAR'));
     }
 
-    public function testInsert()
+    public function providerInsert()
+    {
+        $row1 = ['id' => 1, 'foo' => 'bar', 'bill' => 'baz'];
+        $row2 = ['id' => 2, 'foo' => 'boop', 'bill' => 'bonk'];
+        $row3 = ['id' => 3, 'foo' => 'barne', 'bill' => 'scs'];
+
+        return [
+            'single row'    => [
+                'sometable',
+                $row1,
+                ['sometable', [$row1]],
+            ],
+            'multiple rows' => [
+                'othertable',
+                [$row1, $row2, $row3],
+                ['othertable', [$row1, $row2, $row3]],
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider providerInsert
+     */
+    public function testInsertCanPerformInserts($table, $data, $expectInsertParams)
     {
         // stub migration
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
@@ -179,95 +190,58 @@ class AbstractMigrationTest extends TestCase
             ->setConstructorArgs([[]])
             ->getMock();
         $adapterStub->expects($this->once())
-                    ->method('bulkinsert');
-
-        $table = new Table('testdb', [], $adapterStub);
+            ->method('bulkinsert')
+            ->with(...$expectInsertParams);
 
         $migrationStub->setAdapter($adapterStub);
-        $migrationStub->insert($table, ['row' => 'value']);
+        $migrationStub->insert($table, $data);
     }
 
-    public function testCreateDatabase()
+    public function testInsertThrowsWithVariableSchema()
     {
-        // stub migration
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('mixed column keys');
 
-        // stub adapter
-        $adapterStub = $this->getMockBuilder('\Phinx\Db\Adapter\PdoAdapter')
-            ->setConstructorArgs([[]])
-            ->getMock();
-        $adapterStub->expects($this->once())
-                    ->method('createDatabase')
-                    ->will($this->returnValue([['0' => 'bar', 'foo' => 'bar']]));
-
-        $migrationStub->setAdapter($adapterStub);
-        $migrationStub->createDatabase('testdb', []);
-    }
-
-    public function testDropDatabase()
-    {
-        // stub migration
-        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
-
-        // stub adapter
-        $adapterStub = $this->getMockBuilder('\Phinx\Db\Adapter\PdoAdapter')
-            ->setConstructorArgs([[]])
-            ->getMock();
-        $adapterStub->expects($this->once())
-                    ->method('dropDatabase')
-                    ->will($this->returnValue([['0' => 'bar', 'foo' => 'bar']]));
-
-        $migrationStub->setAdapter($adapterStub);
-        $migrationStub->dropDatabase('testdb');
-    }
-
-    public function testHasTable()
-    {
-        // stub migration
-        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
-
-        // stub adapter
-        $adapterStub = $this->getMockBuilder('\Phinx\Db\Adapter\PdoAdapter')
-            ->setConstructorArgs([[]])
-            ->getMock();
-        $adapterStub->expects($this->once())
-                    ->method('hasTable')
-                    ->will($this->returnValue(true));
-
-        $migrationStub->setAdapter($adapterStub);
-        $this->assertTrue($migrationStub->hasTable('test_table'));
-    }
-
-    public function testTableMethod()
-    {
-        // stub migration
-        $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
-
-        // stub adapter
-        $adapterStub = $this->getMockBuilder('\Phinx\Db\Adapter\PdoAdapter')
-            ->setConstructorArgs([[]])
-            ->getMock();
-        $migrationStub->setAdapter($adapterStub);
-
-        $this->assertInstanceOf(
-            'Phinx\Db\Table',
-            $migrationStub->table('test_table')
+        $migrationStub->insert(
+            'sometable',
+            [
+                ['id' => 1, 'boo' => 'bar'],
+                ['id' => 2, 'boo' => 'foo'],
+                ['id' => 3, 'different_column' => 'foo'],
+            ]
         );
     }
 
-    public function testDropTableMethod()
+    public function testPrepareAndExecute()
     {
-        // stub migration
         $migrationStub = $this->getMockForAbstractClass('\Phinx\Migration\AbstractMigration', [0]);
 
-        // stub adapter
-        $adapterStub = $this->getMockBuilder('\Phinx\Db\Adapter\PdoAdapter')
-            ->setConstructorArgs([[]])
+        $pdoStub = $this->getMockBuilder(\PDO::class)
+            ->disableOriginalConstructor()
             ->getMock();
-        $adapterStub->expects($this->once())
-                    ->method('dropTable');
 
-        $migrationStub->setAdapter($adapterStub);
-        $migrationStub->dropTable('test_table');
+        $statementStub = $this->getMockBuilder(\PDOStatement::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+
+        $pdoStub->expects($this->once())
+            ->method('prepare')
+            ->with('UPDATE foo SET bar = :bax WHERE id=:id')
+            ->willReturn($statementStub);
+
+        $statementStub->expects($this->once())
+            ->method('execute')
+            ->with(['bax' => 'any', 'id' => 15]);
+
+        $migrationStub->setAdapter(
+            new class ($pdoStub) extends MysqlAdapter {
+                public function __construct(\PDO $connection) { $this->connection = $connection; }
+            }
+        );
+
+        $migrationStub->prepareAndExecute('UPDATE foo SET bar = :bax WHERE id=:id', ['bax' => 'any', 'id' => 15]);
     }
+
 }
